requirements.txt
Django==2.0
gunicorn==20.0.4
fabric3
paramiko